rename addTodo action and addReducer, and file formating
