+++ src/Components/Todo.jsx
@@ -2,8 +2,12 @@ import { identifier } from "@babel/types";
 import React from "react";
 import { useDispatch } from "react-redux";
 import delete_btn from "Assets/images/delete.png";
-import { deleteTodo } from "services/actions/addTodo";
-import { ICON_DELETE, IALT_TEXT_DELETE, ALT_TEXT_DELETE } from "Common/constants";
+import { deleteTodo } from "services/actions/actions";
+import {
+  ICON_DELETE,
+  IALT_TEXT_DELETE,
+  ALT_TEXT_DELETE,
+} from "Common/constants";
 
 export default function Todo({ id, text, createdAt }) {
   const dispatch = useDispatch();
