+++ src/services/actions/actions.js
@@ -1,0 +1,19 @@
+export const addTodo = (props) => {
+  const { text } = props;
+  return {
+    type: "ADD",
+    payload: {
+      text: text,
+    },
+  };
+};
+
+
+export const deleteTodo = (id) => {
+  return {
+    type: "DELETE",
+    payload: {
+      id: id,
+    },
+  };
+};
