+++ src/store.js
@@ -1,5 +1,5 @@
 import { createStore } from "redux";
-import addReducer from "./services/reducers/addReducer";
+import addReducer from "./services/reducers/reducers";
 
 const store = createStore(addReducer)
 export default store;
